+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - `apt-get upgrade` commands will no longer stall on upgrades where services starts are required
+- Added `wget` and `curl` to the list of installed dependencies as some server instances don't come with it preinstalled
 
 ----